example/11_otherservice/example_otherservice_01_cloudtasks/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.9.0
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-cloud-tasks==1.1.0